+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add termination conditions so you can ensure an agent stop after a certain amount of time.
 - Drop VCR in favor of WebMock for testing
+- Fix `name` shadowing the built-in function for agent and function
 
 ## [0.4.0] - 2025-02-15